recovery: support sideload over TCP
Author: Andrew Boie <andrew.p.boie@intel.com>
Author: Constantin Musca <constantin.musca@intel.com>

init_socket_transport() used but not defined anywhere,
fortunately this code also exists in system/core/adb.
Copy it over.

Error message changed to printf() to make it show up
in logs but not in UI as this is expected to happen on
devices that lack USB OTG.

Upstream URL:
https://android-review.googlesource.com/#/c/67025/

Category: Device Enablement
Domain: AOSP-recovery
Upstream-candidate: Yes
Issue: ADEV-21
Change-Id: Ic627d9f52eb9329cd593d383bed5f719599fce5a